+++ ghga_devutil/core/exceptions.py
